Huge update. Changed PaymentControllerLogic to use BigDecimal instead of double, which I should have been using the entire time. Changed everything else accordingly so that it can work (Except for PrintReceipt which did not need changing). Also sorted out testing inconsistencies and re-did tests so that they can work. All tests pass, and so this should be treated as a final product unless any issues are spotted.

+++ com.autovend.software/src/com/autovend/software/AttendantIO.java
@@ -8,6 +8,7 @@
 
 package com.autovend.software;
 
+import java.math.BigDecimal;
 
 /**
  * This interface can be used in testing to simulate attendant interactions in certain use cases.
@@ -25,10 +26,10 @@ public interface AttendantIO {
 	 * Simulates an attendant being informed of a change discrepancy.
 	 * Returns nothing.
 	 * 
-	 * @param changeLeft
+	 * @param bigDecimal
 	 * 				Amount of change left that must be given to customer
 	 */
-	public void changeRemainsNoDenom(double changeLeft);
+	public void changeRemainsNoDenom(BigDecimal changeLeft);
 	
 	/**
 	 * Simulates informing the attendant that a duplicate receipt
